test: bump application version in bats tests
Bump the application versions listed in the bats test assersions, after
the devcontainer was rebuilt with newer package versions.

The package versions installed to the container should be locked in the
future to prevent test failures caused by updates. New package versions
should be introduced to this repository using Dependabot.